requirements.txt
streamlit
pandas
fpdf
Pillow
openai